# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MockVision API
  version: 0.1.0
  description: |
    REST API of a MockVision node, shared by the panel (session cookie) and
    automation (API tokens sent as `Authorization: Bearer mvt_…`). Errors use
    problem+json (RFC 9457). State-changing requests with the session cookie
    must send the `X-MockVision-Request` header; requests with a token need
    not. A token with the read scope may only make GET requests and open the
    WebSocket; tokens are created and revoked from the panel only. Live
    updates come through a single WebSocket at `/api/v1/ws` (see the
    `x-websocket` section).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
security:
  - session: []
  - token: []
...
paths:
...
  /profiles/{vendor}/{model}/versions/{version}:
    parameters:
      - { name: vendor, in: path, required: true, schema: { type: string } }
      - { name: model, in: path, required: true, schema: { type: string } }
      - { name: version, in: path, required: true, schema: { type: string } }
    get:
      tags: [profiles]
      operationId: getProfile
      responses:
        "200": { description: Profile, content: { application/json: { schema: { $ref: "#/components/schemas/ProfileDetail" } } } }
        "404": { $ref: "#/components/responses/Problem" }
...
  /tokens:
    get:
      tags: [tokens]
      operationId: listTokens
      summary: API tokens of the current user (panel session only)
      responses:
        "200":
          description: Tokens, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Token" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [tokens]
      operationId: createToken
      summary: Create an API token (panel session only); the secret is returned once
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenInput" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/CreatedToken" } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /tokens/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [tokens]
      operationId: revokeToken
      summary: Revoke an API token (panel session only)
      responses:
        "204": { description: Revoked }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: Background jobs, newest first (D72, D74)
      parameters:
        - { name: status, in: query, description: A status, active (queued, running, waiting or interrupted) or finished, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [rendition, import, renditions.prepare] } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200": { description: Jobs, content: { application/json: { schema: { $ref: "#/components/schemas/JobPage" } } } }
        "422": { $ref: "#/components/responses/Problem" }
    post:
      tags: [jobs]
      operationId: createJob
      summary: Start a job; for now renditions.prepare, which encodes what the cameras need
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "202": { description: Queued, or the open job doing the same work, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "422": { $ref: "#/components/responses/Problem" }
...
  /audit:
    get:
      tags: [audit]
      operationId: listAudit
      summary: Who changed what, when and from where (kept 90 days)
      parameters:
        - { name: origin, in: query, schema: { type: string, enum: [panel, api, camera, system] } }
        - { name: entity_type, in: query, schema: { type: string } }
        - { name: entity_id, in: query, schema: { type: string } }
        - { name: token_id, in: query, schema: { type: string } }
        - { name: action, in: query, description: Exact action or prefix, such as camera, schema: { type: string } }
        - { name: since, in: query, description: Unix milliseconds or RFC 3339, schema: { type: string } }
        - { name: until, in: query, description: Unix milliseconds or RFC 3339, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200": { description: Entries, newest first, content: { application/json: { schema: { $ref: "#/components/schemas/AuditPage" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras:
    get:
      tags: [cameras]
      operationId: listCameras
      parameters:
        - { name: q, in: query, description: Text in the name, address, MAC, serial or tags, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: profile, in: query, description: Profile ID or ID@version, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
      responses:
        "200":
          description: Cameras
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Camera" } }
    post:
      tags: [cameras]
      operationId: createCamera
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCamera" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/actions/bulk:
    post:
      tags: [cameras]
      operationId: bulkCameras
      summary: Start, stop, restart, clone or delete several cameras; each gets its own result
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkAction" }
      responses:
        "200": { description: Results, content: { application/json: { schema: { $ref: "#/components/schemas/BulkResult" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [cameras]
      operationId: getCamera
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [cameras]
      operationId: updateCamera
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCamera" }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
    delete:
      tags: [cameras]
      operationId: deleteCamera
      responses:
        "204": { description: Deleted }
...
  /cameras/{id}/actions/factory-reset:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [cameras]
      operationId: resetCamera
      description: >
        Restores the camera to its profile (RN-10): parameters, accounts and
        protocols; with scope full, also the profile's factory address. The
        picture is kept. A running camera reboots.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope]
              properties:
                scope: { type: string, enum: [settings, full] }
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/actions/clone:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [cameras]
      operationId: cloneCamera
      description: >
        Creates a camera with the same profile, parameters, accounts,
        protocols, picture and targets, with its own ID, serial, MAC and
        address.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CloneCamera" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/users:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [cameras]
      operationId: setCameraUsers
      description: >
        Replaces the camera's accounts (RN-11: at least one admin). An empty
        password keeps the account's current one. Applied at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [users]
              properties:
                users: { type: array, items: { $ref: "#/components/schemas/CameraUserInput" } }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/protocols:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [cameras]
      operationId: setCameraProtocols
      description: >
        Enables, disables or moves protocols of the camera's profile (RN-04).
        A running camera applies them when it restarts (RN-09).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [protocols]
              properties:
                protocols:
                  type: array
                  items:
                    type: object
                    required: [instance]
                    properties:
                      instance: { type: string }
                      enabled: { type: boolean }
                      port: { type: integer, minimum: 0, maximum: 65535 }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "422": { $ref: "#/components/responses/Problem" }
...
  /assets/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [assets]
      operationId: deleteAsset
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: mv_session
    token:
      type: http
      scheme: bearer
      description: API token created in the panel (Settings, API tokens)
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: Problem
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    TokenInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        scopes: { type: array, items: { type: string, enum: [read, write] }, description: write implies read; none means read }
        expires_in_days: { type: [integer, "null"], minimum: 1, maximum: 3650, description: Null for a token that lasts until revoked }
...